+++ redwood-treehouse-host/src/commonMain/kotlin/app/cash/redwood/treehouse/RealTreehouseApp.kt
@@ -141,7 +141,7 @@ internal class RealTreehouseApp<A : AppService> private constructor(
       serializersModule = spec.serializersModule,
       freshnessChecker = spec.freshnessChecker,
     ) { zipline ->
-      spec.bindServices(zipline)
+      spec.bindServices(this, zipline)
     }
   }