+++ CHANGELOG.md
@@ -20,6 +20,9 @@ Fixed:
 Upgraded:
 - Zipline 1.14.0
 
+Breaking:
+  -`TreehouseApp.Spec.bindServices()` now accepts a `TreehouseApp` parameter.
+
 
 ## [0.12.0] - 2024-06-18
 [0.12.0]: https://github.com/cashapp/redwood/releases/tag/0.12.0